Split the viewer connect-gate pin into its own file
It is the only part of V111's suite that needs the WPF Viewer assembly, which is net10.0-windows. With it separated, the rest of the suite reaches only net10.0 projects and can be exercised outside the Windows-only harness.
